backend/requirements.txt
flask
flask_cors
flask_sqlalchemy
psycopg2
Werkzeug
PyJWT